move crt, getenv(), getauxval() to libc
