+++ include/kstdc/sys/auxv.h
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,35 +29,9 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <jinue/getenv.h>
-#include <stdbool.h>
-#include <stddef.h>
+#ifndef _JINUE_KSTDC_SYS_AUXV_H
+#define _JINUE_KSTDC_SYS_AUXV_H
 
-/* This is set by crt.asm. */
-char **jinue_environ = NULL;
+uint32_t getauxval(int type);
 
-const char *jinue_getenv(const char *name) {
-    if(jinue_environ == NULL) {
-        return NULL;
-    }
-
-    for(char **envvar = jinue_environ; *envvar != NULL; ++envvar) {
-        int idx;
-        bool match = true;
-
-        for(idx = 0; name[idx] != '\0'; ++idx) {
-            /* Special case handled here with the general case: the environment
-             * variable string is shorter than name. */
-            if((*envvar)[idx] != name[idx]) {
-                match = false;
-                break;
-            }
-        }
-
-        if(match && (*envvar)[idx] == '=') {
-            return &(*envvar)[idx + 1];
-        }
-    }
-
-    return NULL;
-}
+#endif
