+++ proc/proc.c
@@ -31,7 +31,6 @@
 
 #include <jinue/elf.h>
 #include <jinue/errno.h>
-#include <jinue/getenv.h>
 #include <jinue/ipc.h>
 #include <jinue/memory.h>
 #include <jinue/syscall.h>
@@ -55,9 +54,9 @@ int fd;
 
 char thread_a_stack[THREAD_STACK_SIZE];
 
-extern char **jinue_environ;
+extern char **environ;
 
-extern const Elf32_auxv_t *jinue_auxvp;
+extern const Elf32_auxv_t *_jinue_libc_auxv;
 
 void thread_a(void) {
     int errno;
@@ -94,7 +93,7 @@ void thread_a(void) {
 }
 
 static bool bool_getenv(const char *name) {
-    const char *value = jinue_getenv(name);
+    const char *value = getenv(name);
 
     if(value == NULL) {
         return false;
@@ -147,7 +146,7 @@ static void dump_environ(void) {
 
     printk("Environment variables:\n");
 
-    for(char **envvar = jinue_environ; *envvar != NULL; ++envvar) {
+    for(char **envvar = environ; *envvar != NULL; ++envvar) {
         printk("    %s\n", *envvar);
     }
 }
@@ -196,7 +195,7 @@ static void dump_auxvec(void) {
 
     printk("Auxiliary vectors:\n");
 
-    for(const Elf32_auxv_t *entry = jinue_auxvp; entry->a_type != AT_NULL; ++entry) {
+    for(const Elf32_auxv_t *entry = _jinue_libc_auxv; entry->a_type != AT_NULL; ++entry) {
         const char *name = auxv_type_name(entry->a_type);
 
         if(name != NULL) {
